refactor: move fundraiser types from components/ to types/
