+++ src/components/fundraiser/Fundraiser.tsx
@@ -1,53 +1,8 @@
 import { Link } from 'gatsby'
 import { GatsbyImage } from 'gatsby-plugin-image'
 import { FC } from 'react'
+import { Fundraiser } from '../../types/fundraiser.d'
 import { FundraiserProgress } from './FundraiserProgress'
-
-export type Fundraiser = {
-  id: string
-  name: string
-  title: string
-  /**
-   * Markdown
-   */
-  body: string
-  /**
-   * Image URLs
-   */
-  gallery: Photo[]
-  /**
-   * Describes the funds allocated for this fundraisers.
-   */
-  allocations: {
-    /**
-     * The date when the allocation was made.
-     */
-    date: Date
-    /**
-     * The allocated amount in EUR
-     */
-    amountEUR: number
-    /**
-     * Describes the purpose the allocated funds will be used for.
-     */
-    purpose: string
-  }[]
-}
-
-export type Photo = {
-  /**
-   * URL to media file
-   */
-  url: string
-  /**
-   * Alternative text
-   */
-  alt: string
-  /**
-   * Image process by sharp
-   */
-  gatsbyImageData: any
-}
 
 export const FundraiserCard: FC<{ fundraiser: Fundraiser }> = ({
   fundraiser,
